Add --isolate option
By default, git-duplicate.py use the original commits as parents
of new commits if the original commit being duplicated points to a
parent that is not being duplicated. This makes sense in scenarios
like rebases. However, in scenarios where we want to _disconnect_ history from the
original commits if they are not being duplicated (consider, erasing
history of a branch behind a certain commit), this needs to be avoided.

--isolate makes sure to only consider commits that are being duplicated
as parents of the resulting commits.